fix(seed): scope demo reset to the demo project, extract orchestration
The seed --no-reset path ran TRUNCATE TABLE events/profiles, which wipes every project's rows from these shared, project_id-keyed tables — a latent footgun once the platform is multi-tenant in production. Replace it with the codebase's scoped-delete idiom, ALTER TABLE <t> DELETE WHERE project_id = ? SETTINGS mutations_sync = 1 (mirrors the compliance erasure path and the Postgres ResetDemoProfiles), so only the demo project's rows are cleared; mutations_sync = 1 blocks until the rows are physically gone so the re-backfill can't race stale survivors.

Also remove a dead twin of the same pattern: runProfiles carried an unreachable TRUNCATE TABLE profiles branch (its only caller passed false).

Move the seed orchestration out of cmd/pug/main.go into a new internal/app/seed package exposing Run(ctx, Options), so seedCmd is just flag-parsing + delegation like every other command, and the reset loop returns errors instead of calling os.Exit from library code. The clear-rows flag is bound directly to opts.NoReset (renamed from the misleading 'truncate') rather than defaulted-and-flipped.
